Master program
This is the master program in charge of:

- Connect to an MQTT Server
- Publish topics and read topics from the MQTT Server
- Serve as the interface program between MQTT Server and Telegram through a message sent through the module
- Publish on a regular basis the status of the devices to ThingSpeak MQTT Server.

It is assumed here that the MOSQUITTO Broker is being used. The ThingSpeakUpdater is in charge to replicate MQTT Topics in the broker, so the broker regularly synchronise MOSQUITTO Broker with the ThingSpeak Broker. This is used as an example. It can be modified to include own needs.